requirements.txt
dotenv
fastapi
pydantic
uvicorn
google-cloud-aiplatform
streamlit
openai
google-cloud-aiplatform